02_웹_0912/FastAPI_part1_new/requirements.txt
fastapi==0.99.0
LibRecommender==1.3.0
mangum==0.17.0
pandas==2.1.0
scikit-learn==1.3.0
tensorflow==2.13.0
torch==2.0.1